Add mechanicals for CEC faceplates with brackets
These are after successful 3D print and fitting the 100mm by 80mm BB2 protoboards.
One is with plain faceplate.  The other has six jack holes in the faceplate.

3D Print notes: The plain faceplate has tight fit clearance.  If re-rendored,
It should be the same as for the six hole faceplate.  The six hole faceplate
has less horizontal clearance (across the 3HP) because the jack holes are so
close to the right edge.  Note that these are prototypes.

Also added drawings / diagrams used in the documentation.

Signed-off-by: Darcy Watkins <darcy@xstreamworship.com>